requirements.txt
# requirements.txt
flask==3.0.0
twilio==8.10.0
google-generativeai==0.3.2
gunicorn==21.2.0